Consolidate some of the memory classes
